Add files via upload
Corrected issues with User removal that was previously unknown to be an issue. This issue was caused when I created functions for each individual task and was unaware it was broken. Added a few company specific things that may not apply to those using this. They should not effect the use of the script as you can only use them when using a parameter with specific values. These are in the notes within the script.